requirements.txt
requests~=2.24.0
discord.py~=1.5.0
pycryptodome~=3.9.8